FTApi-M: (fix): Fixed error when saving 'abv' field data

+++ api/migrations/20230208214022-create-beer.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       abv: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DOUBLE(18, 17),
         allowNull: false
       },
       address: {
